requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.2.0
autopep8==1.7.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
contourpy==1.0.5
cycler==0.11.0
Cython==0.29.32
easydict==1.10
fastapi==0.87.0
filetype==1.2.0
flatbuffers==22.12.6
fonttools==4.37.4
frozenlist==1.3.3
google-auth==2.12.0
google-auth-oauthlib==0.4.6
graphviz==0.20.1
grpcio==1.49.1
h11==0.14.0
humanfriendly==10.0
hydra-core==1.3.0
idna==3.4
ImageHash==4.3.1
imageio==2.22.1
imgaug==0.4.0
joblib==1.2.0
kiwisolver==1.4.4
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
mpmath==1.2.1
multidict==6.0.4
networkx==2.8.7
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.1
omegaconf==2.3.0
onnxruntime==1.13.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
Pillow==9.3.0
pooch==1.6.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pycodestyle==2.9.1
pydantic==1.10.2
PyMatting==1.1.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
pytorchyolo==1.7.2
PyWavelets==1.4.1
PyYAML==6.0
rembg==2.0.30
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
Shapely==1.8.5.post1
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
sympy==1.11.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2022.10.10
toml==0.10.2
torch==1.12.1
torchsummary==1.5.1
torchvision==0.13.1
torchviz==0.0.2
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.20.0
watchdog==2.1.9
Werkzeug==2.2.2
yarl==1.8.2
